undo: Move worktree capture under data.undo
Undo stack references now live under data.undo. The helper that collects
current file bytes, file modes, and index entries still sits at
data.undo_worktree beside unrelated services.

Worktree capture exists to build undo checkpoints and is not a general
repository-data API. Its top-level location hides that relationship and
makes it harder to keep checkpoint internals behind the undo package.

This commit continues forming the package by moving worktree path
discovery, saving file contents as Git objects, and index-entry preparation
under data.undo.

The next commit will update tests for that location before checkpoint
restoration follows the same package boundary.
